fn names closer to utility
